# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/v1/users":{"post":{"operationId":"createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersEntity"}}}}},"tags":["用户"],"security":[{"jwt":[]}]}},"/api/v1/users/{id}":{"put":{"operationId":"updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["用户"],"security":[{"jwt":[]}]},"delete":{"operationId":"deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["用户"],"security":[{"jwt":[]}]},"get":{"operationId":"findOneByUserId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersEntity"}}}}},"tags":["用户"],"security":[{"jwt":[]}]}},"/api/v1/members/all":{"get":{"operationId":"findAllMembers","summary":"","description":"通过projectId查询成员","parameters":[{"name":"project_id","required":true,"in":"header","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["成员管理"]}},"/api/v1/members":{"get":{"operationId":"findMembers","summary":"","description":"分页获取成员 且可筛选","parameters":[{"name":"project_id","required":true,"in":"header","schema":{"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["成员管理"]},"post":{"operationId":"createMember","summary":"","description":"添加成员","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMemberDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["成员管理"]}},"/api/v1/members/{id}":{"put":{"operationId":"updateMember","summary":"","description":"更新成员","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMemberDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["成员管理"]},"delete":{"operationId":"deleteMember","summary":"","description":"删除成员","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["成员管理"]}},"/api/v1/auth/login":{"post":{"operationId":"login","summary":"用户登录","description":"生成token","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["登录"]}},"/api/v1/auth/user":{"get":{"operationId":"getUserinfo","summary":"获取用户信息","description":"获取用户信息","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UsersEntity"}}}}},"tags":["登录"],"security":[{"jwt":[]}]}},"/api/v1/projects/all":{"get":{"operationId":"getAllProjects","summary":"获取所有项目","description":"获取所有项目","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectsEntity"}}}}}},"tags":["项目列表"]}},"/api/v1/projects":{"post":{"operationId":"createProject","summary":"创建项目","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectsEntity"}}}}},"tags":["项目列表"]}},"/api/v1/projects/{id}":{"put":{"operationId":"updateProject","summary":"更新项目","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectsEntity"}}}}},"tags":["项目列表"]},"delete":{"operationId":"deleteProject","summary":"删除项目","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["项目列表"]}},"/api/v1/roles":{"get":{"operationId":"findWithParameters","summary":"","description":"分页获取角色","parameters":[{"name":"page","required":false,"in":"query","description":"页数","example":1,"schema":{"type":"number"}},{"name":"size","required":false,"in":"query","description":"每页数量","example":10,"schema":{"type":"number"}},{"name":"projectRoot","required":false,"in":"query","description":"project_root","example":0,"schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["角色管理"]},"post":{"operationId":"createOneRole","summary":"","description":"新增角色权限","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RolesEntity"}}}}},"tags":["角色管理"]}},"/api/v1/roles/{id}":{"get":{"operationId":"findOneRole","summary":"","description":"获取单个角色权限","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RolesEntity"}}}}},"tags":["角色管理"]},"put":{"operationId":"updateRole","summary":"更新角色权限","description":"通过id 更新角色权限","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["角色管理"]},"delete":{"operationId":"remove","summary":"","description":"通过id 删除角色权限","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["角色管理"]}},"/api/v1/roles/all":{"get":{"operationId":"findAllRoles","summary":"","description":"根据project_id 获取所有角色","parameters":[{"name":"project_id","required":true,"in":"header","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RolesEntity"}}}}}},"tags":["角色管理"]}},"/api/v1/resource/category":{"post":{"operationId":"create","summary":"新增分类信息","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryEntity"}}}}},"tags":["检查分类"]},"get":{"operationId":"findAll","summary":"获取所有分类信息","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryEntity"}}}}}},"tags":["检查分类"]}},"/api/v1/resource/category/{id}":{"get":{"operationId":"findOneCategoryById","summary":"根据id 获取分类信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryEntity"}}}}},"tags":["检查分类"]},"put":{"operationId":"updateCategory","summary":"更新分类信息","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}},{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["检查分类"]},"delete":{"operationId":"removeCategory","summary":"删除分类","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["检查分类"]}},"/api/v1/resource/category/extra/{id}":{"get":{"operationId":"findOneCategoryExtraByCategoryId","summary":"获取分类额外信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryExtraEntity"}}}}},"tags":["检查分类"]},"post":{"operationId":"createCategoryExtra","summary":"新增分类额外信息","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryExtraEntity"}}}}},"tags":["检查分类"]}},"/api/v1/resource/instance":{"post":{"operationId":"createInstance","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInstanceDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstanceEntity"}}}}},"tags":["检查实例"],"security":[{"jwt":[]}]},"get":{"operationId":"findInstances","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"size","required":false,"in":"query","schema":{"type":"number"}},{"name":"name","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["检查实例"],"security":[{"jwt":[]}]}},"/api/v1/resource/instance/{id}":{"get":{"operationId":"findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InstanceEntity"}}}}},"tags":["检查实例"],"security":[{"jwt":[]}]},"put":{"operationId":"updateInstance","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateInstanceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["检查实例"],"security":[{"jwt":[]}]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["检查实例"],"security":[{"jwt":[]}]}},"/api/v1/resource/terms":{"post":{"operationId":"create","parameters":[{"name":"project_id","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTermDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TermsEntity"}}}}},"tags":["检查项"]},"get":{"operationId":"getAllTerms","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TermsEntity"}}}}}},"tags":["检查项"]}},"/api/v1/resource/terms/{id}":{"put":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTermDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["检查项"]},"delete":{"operationId":"removeTerm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["检查项"]}},"/api/v1/resource/instance/{instance_id}/items":{"post":{"operationId":"createInstanceItem","parameters":[{"name":"instance_id","required":true,"in":"path","schema":{"type":"string"}},{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InstanceItemsEntity"}}}}}},"tags":["检查实例详情"]},"get":{"operationId":"findAllByInstanceId","parameters":[{"name":"instanceId","required":true,"in":"path","schema":{"type":"string"}},{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InstanceItemsEntity"}}}}}},"tags":["检查实例详情"]}},"/api/v1/resource/instance/{instance_id}/items/unity":{"get":{"operationId":"findByInstanceIdToUnity","parameters":[{"name":"instanceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InstanceItemsEntity"}}}}}},"tags":["检查实例详情"]}},"/api/v1/resource/instance/{instance_id}/items/{id}":{"put":{"operationId":"updateItem","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateItemDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["检查实例详情"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["检查实例详情"]}},"/api/v1/packages":{"get":{"operationId":"findPackages","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPackageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/v1/tasks":{"post":{"operationId":"createTask","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TasksEntity"}}}}},"tags":["任务"]},"get":{"operationId":"findTask","parameters":[{"name":"projectId","required":true,"in":"header","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetTaskDto"}}}},"responses":{"200":{"description":""}},"tags":["任务"]}},"/api/v1/tasks/{id}":{"get":{"operationId":"findOneById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TasksEntity"}}}}},"tags":["任务"]},"patch":{"operationId":"updateTask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["任务"]},"delete":{"operationId":"removeTask","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["任务"]}},"/api/v1/tasks/{taskId}/builds":{"post":{"operationId":"createBuild","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBuildDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildEntity"}}}}},"tags":["构建 builds"]},"get":{"operationId":"findBuilds","parameters":[{"name":"taskId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","description":"page","example":1,"schema":{"type":"number"}},{"name":"size","required":true,"in":"query","description":"size","example":10,"schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BuildEntity"}}}}}},"tags":["构建 builds"]}},"/api/v1/tasks/{taskId}/builds/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildEntity"}}}}},"tags":["构建 builds"]},"put":{"operationId":"updateBuild","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBuildDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuildEntity"}}}}},"tags":["构建 builds"]}}},"info":{"title":"DevOps","description":"DevOps Api Document","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"jwt":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UsersEntity":{"type":"object","properties":{"username":{"type":"string"},"nickname":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"isRoot":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["username","nickname","password","email","isRoot","id","createdAt","updatedAt","deletedAt"]},"UpdateUserDto":{"type":"object","properties":{}},"CreateMemberDto":{"type":"object","properties":{"projectId":{"type":"number"},"roleId":{"type":"number"},"userId":{"type":"number"}},"required":["projectId","roleId","userId"]},"UpdateMemberDto":{"type":"object","properties":{"projectId":{"type":"number"},"roleId":{"type":"number"},"userId":{"type":"number"}}},"LoginAuthDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名","example":"root"},"password":{"type":"string","description":"密码","example":"Zyq@DevOps"}},"required":["username","password"]},"ProjectsEntity":{"type":"object","properties":{"name":{"type":"string"},"imageUrl":{"type":"string"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["name","imageUrl","id","createdAt","updatedAt","deletedAt"]},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdateProjectDto":{"type":"object","properties":{"imageUrl":{"type":"string"},"name":{"type":"string"}}},"RolesEntity":{"type":"object","properties":{"name":{"type":"string"},"privileges":{"type":"string"},"projectRoot":{"type":"number"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["name","privileges","projectRoot","projectId","id","createdAt","updatedAt","deletedAt"]},"CreateRoleDto":{"type":"object","properties":{"privileges":{"type":"object","properties":{}},"name":{"type":"string","description":"name","example":"测试"},"projectRoot":{"type":"number","description":"project_root","example":0}},"required":["privileges","name","projectRoot"]},"UpdateRoleDto":{"type":"object","properties":{"privileges":{"type":"object","properties":{}},"name":{"type":"string","description":"name","example":"测试"},"projectRoot":{"type":"number","description":"project_root","example":0}},"required":["privileges"]},"CreateCategoryDto":{"type":"object","properties":{"categroyId":{"type":"number","description":"分类名称","example":"纹理检查"},"categoryUid":{"type":"string"},"globalParams":{"type":"object"}},"required":["categroyId","categoryUid","globalParams"]},"CategoryEntity":{"type":"object","properties":{"categoryName":{"type":"string"},"categoryUid":{"type":"string"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["categoryName","categoryUid","projectId","id","createdAt","updatedAt","deletedAt"]},"UpdateCategoryDto":{"type":"object","properties":{"categoryName":{"type":"string","description":"Uid","example":"TextureDetector"},"categoryUid":{"type":"string","description":"分类名称","example":"纹理检查"}}},"CategoryExtraEntity":{"type":"object","properties":{"categoryId":{"type":"number"},"categoryUid":{"type":"string"},"globalParams":{"type":"object"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["categoryId","categoryUid","globalParams","projectId","id","createdAt","updatedAt","deletedAt"]},"CreateInstanceDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"InstanceEntity":{"type":"object","properties":{"name":{"type":"string","description":"实例名称","example":"资源检查"},"userId":{"type":"number","description":"创建者","example":1},"projectId":{"type":"number","description":"项目id","example":1},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["name","userId","projectId","id","createdAt","updatedAt","deletedAt"]},"UpdateInstanceDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateTermDto":{"type":"object","properties":{"categoryId":{"type":"number","description":"检查分类id","example":1},"ruleUid":{"type":"string","description":"rule_uid","example":"rule_uid"},"ruleName":{"type":"string","description":"rule_name","example":"rule_name"},"ruleDesc":{"type":"string","description":"rule_desc","example":"rule_desc"},"suggest":{"type":"string","description":"suggest","example":"suggest"},"level":{"type":"number","description":"level","example":1},"detectPaths":{"type":"string","description":"detect_paths","example":"detect_paths"},"filterPaths":{"type":"string","description":"filter_paths","example":"filter_paths"},"filterRegex":{"type":"string","description":"filter_regex","example":"filter_regex"},"thresholdRange":{"type":"string","description":"threshold_range","example":"threshold_range"},"thresholdValue":{"type":"string","description":"threshold_value","example":"threshold_value"},"enabled":{"type":"number","description":"enabled","example":1}},"required":["categoryId","ruleUid","ruleName","ruleDesc","level","detectPaths","enabled"]},"TermsEntity":{"type":"object","properties":{"categoryId":{"type":"number"},"ruleUid":{"type":"string"},"ruleName":{"type":"string"},"ruleDesc":{"type":"string"},"suggest":{"type":"string"},"level":{"type":"number"},"detectPaths":{"type":"string"},"filterPaths":{"type":"string"},"filterRegex":{"type":"string"},"thresholdRange":{"type":"string"},"thresholdValue":{"type":"string"},"enabled":{"type":"number"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["categoryId","ruleUid","ruleName","ruleDesc","suggest","level","detectPaths","filterPaths","filterRegex","thresholdRange","thresholdValue","enabled","projectId","id","createdAt","updatedAt","deletedAt"]},"UpdateTermDto":{"type":"object","properties":{}},"CreateItemDto":{"type":"object","properties":{"ids":{"description":"实例id","example":[1],"type":"array","items":{"type":"number"}}},"required":["ids"]},"InstanceItemsEntity":{"type":"object","properties":{"instanceId":{"type":"number"},"termId":{"type":"number"},"detectPaths":{"type":"string"},"filterPaths":{"type":"string"},"filterRegex":{"type":"string"},"thresholdValue":{"type":"string"},"enabled":{"type":"number"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["instanceId","termId","detectPaths","filterPaths","filterRegex","thresholdValue","enabled","projectId","id","createdAt","updatedAt","deletedAt"]},"UpdateItemDto":{"type":"object","properties":{"enabled":{"type":"number","description":"是否开启","example":1},"thresholdValue":{"type":"string","description":"阈值","example":"1"},"filterRegex":{"type":"string","description":"过滤正则","example":"/^[0-9]*$/"},"filterPaths":{"type":"string","description":"过滤路径","example":"/a/b,/c/d"},"detectPaths":{"type":"string","description":"检测目录","example":"/a/b,/c/d"}}},"GetPackageDto":{"type":"object","properties":{"page":{"type":"number"},"size":{"type":"number"}},"required":["page","size"]},"CreateTaskDto":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"description":{"type":"string"},"documentUrl":{"type":"string"},"jenkinsId":{"type":"number"},"viewId":{"type":"number"}},"required":["name","displayName","description","documentUrl","jenkinsId","viewId"]},"TasksEntity":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"description":{"type":"string"},"documentUrl":{"type":"string"},"jenkinsId":{"type":"number"},"viewId":{"type":"number"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["name","displayName","description","documentUrl","jenkinsId","viewId","projectId","id","createdAt","updatedAt","deletedAt"]},"GetTaskDto":{"type":"object","properties":{"page":{"type":"number","description":"页数","example":1},"size":{"type":"number","description":"每页数量","example":10}},"required":["page","size"]},"UpdateTaskDto":{"type":"object","properties":{"name":{"type":"string"},"displayName":{"type":"string"},"description":{"type":"string"},"documentUrl":{"type":"string"},"jenkinsId":{"type":"number"},"viewId":{"type":"number"}}},"CreateBuildDto":{"type":"object","properties":{"taskId":{"type":"number","description":"任务id","example":1},"status":{"type":"number","description":"任务状态","example":1},"userId":{"type":"number","description":"用户id","example":1},"jobName":{"type":"string","description":"job_name","example":"android_test"},"buildType":{"type":"string","description":"build_type","example":"resource"}},"required":["taskId","status","userId","jobName","buildType"]},"BuildEntity":{"type":"object","properties":{"taskId":{"type":"number"},"status":{"type":"number"},"userId":{"type":"number"},"jobName":{"type":"string"},"number":{"type":"number"},"parameters":{"type":"object"},"buildType":{"type":"string"},"fileId":{"type":"string"},"resultId":{"type":"number"},"projectId":{"type":"number"},"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"}},"required":["taskId","status","userId","jobName","number","parameters","buildType","fileId","resultId","projectId","id","createdAt","updatedAt","deletedAt"]},"UpdateBuildDto":{"type":"object","properties":{"taskId":{"type":"number","description":"任务id","example":1},"status":{"type":"number","description":"任务状态","example":1},"userId":{"type":"number","description":"用户id","example":1},"jobName":{"type":"string","description":"job_name","example":"android_test"},"buildType":{"type":"string","description":"build_type","example":"resource"}}}}}}